% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{monitor}
\alias{monitor}
\title{Monitor files for changes}
\usage{
monitor(
  dir = getwd(),
  file,
  ext = "R",
  ignore = NULL,
  delay = 1,
  log_file = NULL,
  clear_log_file = FALSE
)
}
\arguments{
\item{dir}{Character vector. Directory to monitor for changes. Can be
multiple. Defaults to the directory from where this function is called from.}

\item{file}{String, file path. File to rerun when changes are detected in
\code{dir}.}

\item{ext}{Character vector. Extensions to watch. Defaults to \code{.R} files.}

\item{ignore}{Character vector. Files or directories to ignore. Default is
\code{NULL}.}

\item{delay}{Length one numeric. Number of seconds to wait before checking
for file changes again. Defaults to \code{1}.}

\item{log_file}{String, file path. Log file. Defaults to \code{NULL}.}

\item{clear_log_file}{Logical. Clear log file on each change? Defaults to
\code{FALSE}.}
}
\description{
Monitor files for changes
}
\examples{
\dontrun{
rmon::monitor(
  dir = ".",
  file = "app.R",
  ignore = c("dev.R", "test")
)
}
}
